third_party/WebKit/LayoutTests/media/encrypted-media/encrypted-media-playback-expected.txt

Test playback of encrypted media using Clear Key key system.

EVENT(needkey)
EVENT(needkey)
EVENT(message)
EVENT(ready)
EVENT(playing)
This is a testharness.js-based test.
PASS Playback using Clear Key key system. 
Harness: the test ran to completion.
